Refactor baby_city_za
